requirements.txt
streamlit==1.26.0
PyPDF2==3.0.1
langchain==0.95.0
langchain-google-genai==0.0.4
google-generativeai==0.1.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
